gridNode restructuring
Changed the gridNode class so that there are no inherited types. Instead, a serial number is used to refer to all data within the SQlite database and type-specific functions can be called with checks made on the type of the component. The Catalog now holds UniqueIDs for reference to the SQLite database, and when double clicked a gridNode is created which will serve as the instance. This gridNode will later hold connection data which can be used in the network editing window and grid solving algorithms.